extract node connection and drag files from editor
