+++ app/javascript/editor/node_editor.js
@@ -1,5 +1,7 @@
 import EditorApi from "editor/api";
 import NodeFormHandler from "editor/form_handler";
+import ConnectionManager from "editor/connection_manager";
+import DragManager from "editor/drag_manager";
 
 class NodeEditor {
   constructor(canvasId) {
@@ -10,16 +12,19 @@ class NodeEditor {
     this.api = new EditorApi(this.botId);
 
     this.nodes = new Map();
-    this.connections = [];
-    this.selectedNode = null;
     this.currentTool = 'select';
-    this.isDragging = false;
-    this.positionChanged = false;
-    this.isConnecting = false;
-    this.connectSource = null;
-    this.dragOffset = { x: 0, y: 0 };
     
     this.formHandler = new NodeFormHandler(this.api, this.nodes);
+    this.connectionManager = new ConnectionManager(this.api, this.nodes);
+    this.dragManager = new DragManager(this.nodes, this.nodesCanvas, this.api, this.connectionManager);
+    
+    // Set up drag callbacks
+    this.dragManager.setCallbacks({
+      onDragStart: (nodeId) => {
+        const node = this.nodes.get(nodeId);
+        this.openEditor(nodeId);
+      }
+    });
 
     this.init();
   }
@@ -27,7 +32,6 @@ class NodeEditor {
   init() {
     this.loadNodes();
     this.setupEventListeners();
-    this.setupPaletteDrag();
   }
 
   loadNodes() {
@@ -42,60 +46,8 @@ class NodeEditor {
         }
       });
     });
-    this.loadConnections();
-    this.drawExistingConnections();
-  }
-
-  ensureNodeSpacing() {
-    const MIN_DISTANCE = 120;
-    const nodesArray = Array.from(this.nodes.values());
-    
-    for (let i = 0; i < nodesArray.length; i++) {
-      for (let j = i + 1; j < nodesArray.length; j++) {
-        const nodeA = nodesArray[i];
-        const nodeB = nodesArray[j];
-        
-        const dx = nodeA.position.x - nodeB.position.x;
-        const dy = nodeA.position.y - nodeB.position.y;
-        const distance = Math.sqrt(dx * dx + dy * dy);
-        
-        if (distance < MIN_DISTANCE) {
-          nodeB.position.x = nodeA.position.x + MIN_DISTANCE;
-          nodeB.position.y = nodeA.position.y;
-          nodeB.element.style.left = `${nodeB.position.x}px`;
-          nodeB.element.style.top = `${nodeB.position.y}px`;
-          
-          this.saveNodePosition(parseInt(nodeB.element.dataset.id), nodeB.position.x, nodeB.position.y);
-        }
-      }
-    }
-  }
-
-  drawExistingConnections() {
-    const svg = document.getElementById('connections-canvas');
-    const connectionsData = svg?.dataset.connections;
-    if (!connectionsData) return;
-    
-    try {
-      const connections = JSON.parse(connectionsData);
-      connections.forEach(conn => {
-        if (this.nodes.has(conn.source) && this.nodes.has(conn.target)) {
-          this.drawConnection(conn.source, conn.target, conn.id);
-        }
-      });
-    } catch (e) {
-      console.error('Failed to parse connections:', e);
-    }
-  }
-
-  loadConnections() {
-    this.nodes.forEach((node, id) => {
-      const output = node.element.querySelector('.output');
-      
-      if (output) {
-        output.addEventListener('mousedown', (e) => this.startConnection(e, id));
-      }
-    });
+    this.connectionManager.loadConnections();
+    this.connectionManager.drawExistingConnections();
   }
 
   setupEventListeners() {
@@ -113,17 +65,6 @@ class NodeEditor {
     document.addEventListener('mousedown', (e) => this.handleMouseDown(e));
     document.addEventListener('mousemove', (e) => this.handleMouseMove(e));
     document.addEventListener('mouseup', (e) => this.handleMouseUp(e));
-  }
-
-  setupPaletteDrag() {
-    document.querySelectorAll('.palette-item').forEach(item => {
-      item.addEventListener('dragstart', (e) => {
-        e.dataTransfer.setData('nodeType', item.dataset.type);
-      });
-    });
-
-    this.nodesCanvas.addEventListener('dragover', (e) => e.preventDefault());
-    this.nodesCanvas.addEventListener('drop', (e) => this.handleDrop(e));
   }
 
   setTool(tool) {
@@ -149,255 +90,41 @@ class NodeEditor {
 
     if (this.currentTool === 'select' && nodeEl) {
       const clickedNodeId = parseInt(nodeEl.dataset.id);
-      this.selectedNode = clickedNodeId;
-      this.isDragging = true;
-      this.positionChanged = false;
-      const rect = nodeEl.getBoundingClientRect();
-      const canvasRect = this.nodesCanvas.getBoundingClientRect();
-      this.dragOffset = {
-        x: e.clientX - rect.left,
-        y: e.clientY - rect.top
-      };
-      
-      const node = this.nodes.get(clickedNodeId);
-      this.openEditor(clickedNodeId);
-      this.showPaletteForNode(node.type);
-      
+      this.dragManager.startDrag(e, clickedNodeId);
       e.preventDefault();
     } else if (!nodeEl) {
-      this.hideAllPalettes();
       this.closeEditor();
-      this.selectedNode = null;
-    }
-  }
-
-  hideAllPalettes() {
-    document.querySelectorAll('.palette-section').forEach(section => {
-      section.classList.remove('visible');
-    });
-  }
-
-  showPaletteForNode(nodeType) {
-    this.hideAllPalettes();
-    if (nodeType === 'condition') {
-      const sections = document.querySelectorAll('.palette-section');
-      if (sections[0]) sections[0].classList.add('visible');
-    } else if (nodeType === 'action') {
-      const sections = document.querySelectorAll('.palette-section');
-      if (sections[1]) sections[1].classList.add('visible');
     }
   }
 
   handleMouseMove(e) {
-    if (this.isDragging && this.selectedNode) {
-      const node = this.nodes.get(this.selectedNode);
-      const canvasRect = this.nodesCanvas.getBoundingClientRect();
-      
-      node.position.x = e.clientX - canvasRect.left - this.dragOffset.x;
-      node.position.y = e.clientY - canvasRect.top - this.dragOffset.y;
-      
-      node.element.style.left = `${node.position.x}px`;
-      node.element.style.top = `${node.position.y}px`;
-      
-      this.positionChanged = true;
-      
-      this.updateConnections();
-    }
-
-    if (this.isConnecting) {
-      this.updateConnectionLine(e.clientX, e.clientY);
+    this.dragManager.handleMouseMove(e);
+    
+    if (this.connectionManager.isConnecting) {
+      this.connectionManager.updateConnectionLine(e.clientX, e.clientY);
     }
   }
 
   handleMouseUp(e) {
-    if (this.isDragging && this.selectedNode) {
-      const node = this.nodes.get(this.selectedNode);
-      if (this.positionChanged) {
-        this.saveNodePosition(this.selectedNode, node.position.x, node.position.y);
-      }
-    }
-    this.isDragging = false;
-    this.positionChanged = false;
-    this.selectedNode = null;
+    this.dragManager.handleMouseUp();
 
-    if (this.isConnecting) {
+    if (this.connectionManager.isConnecting) {
       const inputConnector = e.target.closest('.node-connector.input');
       if (inputConnector) {
         const targetNode = inputConnector.closest('.node');
-        if (targetNode && this.connectSource) {
-          const sourceNode = this.nodes.get(this.connectSource.nodeId);
-          if (sourceNode && parseInt(targetNode.dataset.id) !== this.connectSource.nodeId) {
-            this.createConnection(this.connectSource.nodeId, parseInt(targetNode.dataset.id));
+        if (targetNode && this.connectionManager.connectSource) {
+          const sourceNode = this.nodes.get(this.connectionManager.connectSource.nodeId);
+          if (sourceNode && parseInt(targetNode.dataset.id) !== this.connectionManager.connectSource.nodeId) {
+            this.connectionManager.createConnection(this.connectionManager.connectSource.nodeId, parseInt(targetNode.dataset.id));
           }
         }
       }
-      this.endConnection();
+      this.connectionManager.endConnection();
     }
   }
 
   startConnection(e, nodeId) {
-    this.isConnecting = true;
-    this.connectSource = { nodeId };
-    e.stopPropagation();
-    e.preventDefault();
-    
-    this.createTempLine(e.clientX, e.clientY);
-  }
-
-  createTempLine(x, y) {
-    const svg = document.getElementById('connections-canvas');
-    this.tempLine = document.createElementNS('http://www.w3.org/2000/svg', 'line');
-    const sourceNode = this.nodes.get(this.connectSource.nodeId);
-    const sourceEl = sourceNode.element;
-    const nodesCanvasRect = this.nodesCanvas.getBoundingClientRect();
-    
-    const startX = parseFloat(sourceEl.style.left) + 100;
-    const startY = parseFloat(sourceEl.style.top) + 30;
-    
-    this.tempLine.setAttribute('x1', startX);
-    this.tempLine.setAttribute('y1', startY);
-    this.tempLine.setAttribute('x2', x - nodesCanvasRect.left);
-    this.tempLine.setAttribute('y2', y - nodesCanvasRect.top);
-    this.tempLine.setAttribute('stroke', '#4CAF50');
-    this.tempLine.setAttribute('stroke-width', '3');
-    this.tempLine.setAttribute('stroke-dasharray', '5,5');
-    
-    svg.appendChild(this.tempLine);
-  }
-
-  updateConnectionLine(x, y) {
-    if (!this.tempLine) return;
-    
-    const nodesCanvasRect = this.nodesCanvas.getBoundingClientRect();
-    this.tempLine.setAttribute('x2', x - nodesCanvasRect.left);
-    this.tempLine.setAttribute('y2', y - nodesCanvasRect.top);
-  }
-
-  endConnection() {
-    if (this.tempLine) {
-      this.tempLine.remove();
-      this.tempLine = null;
-    }
-    this.isConnecting = false;
-    this.connectSource = null;
-  }
-  
-  createConnection(sourceId, targetId) {
-    if (!this.botId) return;
-    
-    this.api.createConnection(sourceId, targetId)
-    .then(conn => {
-      this.drawConnection(sourceId, targetId, conn.id);
-    })
-    .catch(err => console.error('Connection failed:', err));
-  }
-
-  drawConnection(sourceId, targetId, connectionId) {
-    const svg = document.getElementById('connections-canvas');
-    const sourceNode = this.nodes.get(sourceId);
-    const targetNode = this.nodes.get(targetId);
-    
-    const line = document.createElementNS('http://www.w3.org/2000/svg', 'line');
-    const startX = parseFloat(sourceNode.element.style.left) + 100;
-    const startY = parseFloat(sourceNode.element.style.top) + 30;
-    const endX = parseFloat(targetNode.element.style.left);
-    const endY = parseFloat(targetNode.element.style.top) + 30;
-    
-    const hitArea = document.createElementNS('http://www.w3.org/2000/svg', 'line');
-    hitArea.setAttribute('x1', startX);
-    hitArea.setAttribute('y1', startY);
-    hitArea.setAttribute('x2', endX);
-    hitArea.setAttribute('y2', endY);
-    hitArea.setAttribute('stroke', 'transparent');
-    hitArea.setAttribute('stroke-width', '20');
-    hitArea.style.pointerEvents = 'stroke';
-    hitArea.dataset.sourceId = sourceId;
-    hitArea.dataset.targetId = targetId;
-    hitArea.dataset.connectionId = connectionId;
-    
-    line.setAttribute('x1', startX);
-    line.setAttribute('y1', startY);
-    line.setAttribute('x2', endX);
-    line.setAttribute('y2', endY);
-    line.setAttribute('stroke', '#4CAF50');
-    line.setAttribute('stroke-width', '2');
-    line.style.pointerEvents = 'none';
-    line.dataset.sourceId = sourceId;
-    line.dataset.targetId = targetId;
-    line.dataset.connectionId = connectionId;
-    
-    const midX = (startX + endX) / 2;
-    const midY = (startY + endY) / 2;
-    
-    const deleteBtn = document.createElement('button');
-    deleteBtn.className = 'connection-delete-btn';
-    deleteBtn.textContent = '×';
-    deleteBtn.style.left = `${midX}px`;
-    deleteBtn.style.top = `${midY}px`;
-    deleteBtn.dataset.sourceId = sourceId;
-    deleteBtn.dataset.targetId = targetId;
-    deleteBtn.dataset.connectionId = connectionId;
-    
-    hitArea.addEventListener('mouseenter', (e) => {
-      if (!this.nodes.has(sourceId) || !this.nodes.has(targetId)) return;
-      const sourceEl = this.nodes.get(sourceId).element;
-      const targetEl = this.nodes.get(targetId).element;
-      if (sourceEl.matches(':hover') || targetEl.matches(':hover')) return;
-      deleteBtn.style.display = 'block';
-    });
-    
-    hitArea.addEventListener('mouseleave', () => {
-      deleteBtn.style.display = 'none';
-    });
-    
-    deleteBtn.addEventListener('mouseenter', () => {
-      deleteBtn.style.display = 'block';
-    });
-    
-    deleteBtn.addEventListener('mouseleave', () => {
-      deleteBtn.style.display = 'none';
-    });
-    
-    deleteBtn.addEventListener('click', (e) => {
-      e.stopPropagation();
-      this.disconnectNode(sourceId, targetId);
-    });
-    
-    svg.appendChild(hitArea);
-    svg.appendChild(line);
-    this.nodesCanvas.appendChild(deleteBtn);
-  }
-
-  updateConnections() {
-    const svg = document.getElementById('connections-canvas');
-    svg.querySelectorAll('line').forEach(line => {
-      const sourceId = parseInt(line.dataset.sourceId);
-      const targetId = parseInt(line.dataset.targetId);
-      
-      const sourceNode = this.nodes.get(sourceId);
-      const targetNode = this.nodes.get(targetId);
-      
-      if (!sourceNode || !targetNode) return;
-      
-      const startX = parseFloat(sourceNode.element.style.left) + 100;
-      const startY = parseFloat(sourceNode.element.style.top) + 30;
-      const endX = parseFloat(targetNode.element.style.left);
-      const endY = parseFloat(targetNode.element.style.top) + 30;
-      
-      line.setAttribute('x1', startX);
-      line.setAttribute('y1', startY);
-      line.setAttribute('x2', endX);
-      line.setAttribute('y2', endY);
-      
-      const midX = (startX + endX) / 2;
-      const midY = (startY + endY) / 2;
-      
-      const deleteBtn = document.querySelector('.connection-delete-btn[data-source-id="' + sourceId + '"][data-target-id="' + targetId + '"]');
-      if (deleteBtn) {
-        deleteBtn.style.left = midX + 'px';
-        deleteBtn.style.top = midY + 'px';
-      }
-    });
+    this.connectionManager.startConnection(e, nodeId);
   }
 
   renderNode(node) {
@@ -425,7 +152,7 @@ class NodeEditor {
       position: { x: node.position_x, y: node.position_y }
     });
     
-    this.loadConnections();
+    this.connectionManager.loadConnections();
   }
 
   deleteNode(nodeId) {
@@ -472,13 +199,6 @@ class NodeEditor {
 
   closeEditor() {
     this.formHandler.closeEditor();
-  }
-
-  saveNodePosition(nodeId, x, y) {
-    if (!this.botId) return;
-    
-    this.api.updateNodePosition(nodeId, x, y)
-    .catch(err => console.error('Failed to update position:', err));
   }
 
   addNode(type) {
